Put the project explorer in its own python file and used links in the Widget notebook to link widgets through the new elements_by_id traitlet
