+++ src/pymbe/api.py
@@ -1,2 +1,3 @@
-from .widget.client import SysML2ClientWidget, ElementDetails
+from .widget.client import SysML2ClientWidget
+from .widget.explorer import ProjectExplorer
 from .widget.graph import SysML2LPGWidget
